WIP - Incomplete
- Add 404Controller.
- Add healthCheckController.
- Add photosController.
- Add photo model.
- Update server to consume the module extractions.

Took 2 hour 31 minutes
