+++ src/controllers/Erorr400Controller.js
@@ -1,0 +1,10 @@
+/*
+ * Joe Ivans, student, Code Fellows Seattle
+ */
+
+'use strict';
+
+module.exports =
+  async function getNotFound(request, response) {
+    response.status(404).send('I can\'t find it.');
+  };
